save mixtape details
