+++ frontend/src/components/Container/Container.module.css
@@ -14,8 +14,7 @@ h1 {
 }
 .body {
   background-image: linear-gradient(#000000, #1a0f37);
-  width: 100vw;
-  height: 100vh;
+  min-height: 100vh;
   margin: 0;
   display: flex;
 }
@@ -23,8 +22,8 @@ h1 {
   position: absolute;
   top: 0;
   left: 0;
-  right: 0;
-  bottom: 0;
+  height: 100vh;
+  width: 100%;
   background-color: #fff;
   opacity: 6%;
 }
