+++ frontend/src/pages/DesignCassette/containers/ControlPanel/ControlPanel.css
@@ -1,8 +1,4 @@
 .ControlPanel {
-  position: absolute;
-  top: 10px;
-  left: 10px;
-
   box-sizing: border-box;
   border: 1px solid black;
   background-color: white;
