+++ frontend/src/pages/DesignCassette/containers/Workspace/Workspace.css
@@ -1,14 +1,12 @@
 .Workspace {
-  position: absolute;
-  top: 10px;
-  left: 310px;
+  position: relative;
 
   box-sizing: border-box;
   border: 1px solid black;
   border-left-width: 0px;
   background-color: white;
 
-  width: 800px;
+  width: 600px;
   height: 700px;
 
   display: flex;
